fix connection string for mlflow
