+++ .pre-commit-config.yaml
@@ -1,7 +1,7 @@
 exclude: (^.idea/)
 repos:
   - repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v4.6.0
+    rev: v5.0.0
     hooks:
       - id: end-of-file-fixer
         types: [ python ]
